Plugin mechanisms for redfish objects management
