+++ tests/conf.json
@@ -8,12 +8,16 @@
         "fs_root": "Resources",
         "subscribers_root": "EventService/Subscriptions"
     },
-    "event_handler": {
-        "module_name": "event_handlers.redfish.redfish_event_handler",
+    "events_handler": {
+        "module_name": "events_handlers.redfish.redfish_event_handler",
         "class_name": "RedfishEventHandler"
     },
     "handlers": {
         "subscription_handler": "redfish",
         "event_handler": "redfish"
+    },
+    "objects_handler": {
+        "module_name": "objects_handlers.sunfish_server.redfish_object_handler",
+        "class_name": "RedfishObjectHandler"
     }
 }
